#5: Reject wrong GitHub tokens
