+++ app/jobs/load_events_job.rb
@@ -2,6 +2,6 @@ class LoadEventsJob < ApplicationJob
   queue_as :default
 
   def perform(token, username)
-    LoadEventsFromGithubService.new(token, username).perform
+    GithubActions::LoadEventsService.new(token, username).perform
   end
 end
